parser: decouple parsers
make it more convenient for others to build more complex liquid tags
